Fix some NSUInteger / int mismatches.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@7782 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Headers/ETContainer.h
@@ -75,10 +75,10 @@
 /* Layout Item Tree */
 
 - (void) addItem: (ETLayoutItem *)item;
-- (void) insertItem: (ETLayoutItem *)item atIndex: (int)index;
+- (void) insertItem: (ETLayoutItem *)item atIndex: (NSUInteger)index;
 - (void) removeItem: (ETLayoutItem *)item;
-- (void) removeItemAtIndex: (int)index;
-- (ETLayoutItem *) itemAtIndex: (int)index;
+- (void) removeItemAtIndex: (NSUInteger)index;
+- (ETLayoutItem *) itemAtIndex: (NSUInteger)index;
 - (int) indexOfItem: (ETLayoutItem *)item;
 - (BOOL) containsItem: (ETLayoutItem *)item;
 - (int) numberOfItems;
